fix: cleanup in circulating supply
move run_circulating_supply_computation to mod rs
prepare the place for other aggregated modules
get rid of cargo clippy warnings
