+++ src/main.rs
@@ -11,7 +11,6 @@ use tokio::sync::mpsc;
 use tracing::info;
 use tracing_subscriber::EnvFilter;
 
-use crate::aggregated::circulating_supply::circulating_supply_provider;
 use crate::configs::{Opts, SubCommand};
 
 mod aggregated;
@@ -277,16 +276,7 @@ fn main() {
                 ));
 
                 // Spawning the computation of aggregated data
-                let view_client = indexer.client_actors().0;
-
-                if indexer.near_config().genesis.config.chain_id == "mainnet" {
-                    actix::spawn(
-                        circulating_supply_provider::run_circulating_supply_computation(
-                            view_client,
-                            pool,
-                        ),
-                    );
-                }
+                aggregated::spawn_aggregated_computations(pool, &indexer);
             });
             system.run().unwrap();
         }
